Reorganize and de-relativize includes for test utilites
